Small Business - DESKTOP: Sync stamp and hallmark.                   Move the logo so there is 30px to left of the logo and 30px below the logo (based on the H, in Human Energy)
TABLET:
Move the logo so there is 30px to left of the logo and 30px below the logo (based on the H, in Human Energy)
Please use the Learn More CTA from the desktop version, which is larger and better for tablet.
In the portrait version could we have the man shifted over to the left about 40px?
Please ensure the layout of the tablet is consistent for all tablet screen sizes. On a large tablet screen, the copy runs very close to the man's ear.
 smoothing the animation for the INLINE units pertains to the motion of the man and the background. Additionally, could we slow down the animation for the Inline desktop and tablet units.